Bump spec version to v0.4.0
