+++ Godeps/Godeps.json
@@ -57,9 +57,9 @@
 			"Rev": "f7137ae6b19afbfd61a94b746fda3b3fe0491874"
 		},
 		{
-			"ImportPath": "github.com/opencontainers/specs",
-			"Comment": "v0.3.0-15-ga1e32a8",
-			"Rev": "a1e32a8ead2ba57adce3e36e956b4dc32c1b85c4"
+			"ImportPath": "github.com/opencontainers/specs/specs-go",
+			"Comment": "v0.4.0",
+			"Rev": "3ce138b1934bf227a418e241ead496c383eaba1c"
 		},
 		{
 			"ImportPath": "github.com/seccomp/libseccomp-golang",
