+++ Godeps/_workspace/src/github.com/opencontainers/specs/README.md
@@ -5,15 +5,18 @@
 
 Table of Contents
 
-- [Container Principles](principles.md)
-- [Specification Style](style.md)
+- [Introduction](README.md)
+  - [Code of Conduct](code-of-conduct.md)
+  - [Container Principles](principles.md)
+  - [Style and Conventions](style.md)
+  - [Roadmap](ROADMAP.md)
+  - [Implementations](implementations.md)
 - [Filesystem Bundle](bundle.md)
+- [Runtime and Lifecycle](runtime.md)
+  - [Linux Specific Runtime](runtime-linux.md)
 - Configuration
   - [General](config.md)
   - [Linux-specific](config-linux.md)
-- [Runtime and Lifecycle](runtime.md)
-  - [Linux Specific Runtime](runtime-linux.md)
-- [Implementations](implementations.md)
 - [Glossary](glossary.md)
 
 In the specifications in the above table of contents, the keywords "MUST", "MUST NOT", "REQUIRED", "SHALL", "SHALL NOT", "SHOULD", "SHOULD NOT", "RECOMMENDED", "MAY", and "OPTIONAL" are to be interpreted as described in [RFC 2119](http://tools.ietf.org/html/rfc2119) (Bradner, S., "Key words for use in RFCs to Indicate Requirement Levels", BCP 14, RFC 2119, March 1997).
@@ -154,5 +157,4 @@ Read more on [How to Write a Git Commit Message](http://chris.beams.io/posts/git
   * If there was important/useful/essential conversation or information, copy or include a reference
 8. When possible, one keyword to scope the change in the subject (i.e. "README: ...", "runtime: ...")
 
-[BlueJeans]: https://bluejeans.com/1771332256/
 [UberConference]: https://www.uberconference.com/ssaul
