+++ Godeps/_workspace/src/github.com/opencontainers/specs/specs-go/state.go
@@ -1,9 +1,6 @@
 package specs
 
 // State holds information about the runtime state of the container.
-// This information will be stored in a file called `state.json`.
-// The location of this file will be operating system specific. On Linux
-// it will be in `/run/opencontainers/runc/<containerID>/state.json`
 type State struct {
 	// Version is the version of the specification that is supported.
 	Version string `json:"version"`
